Added UserTest and moved test cases to test folder
